Replication of Illustrator Doc
To the best of my ability so far, I have replicated the index.html page of our illustrator documents.

For our icons, we are using .svgs, enabling us to edit more than we normally would be able to if we used .png or .jpg types.

I have changed some of the names of our files and folders, hopefully that is not a big issue.

There are many empty files which will be added to in the coming weeks, but as of right now, index.html, collect.css, main-design.css, and portfolio-style.css are working as intended.

If time permits, please feel free to review and get back to me with any suggestions. Excited to meet back with you all to discuss this further on Monday!