merge: merged files | style:fixed images
Disregard most of the previous message— I found the commit, so I've merged my new stuff in.

I created uniform names for the avatar photos, renamed the placeholder images appropriately (so we will only need to make one change for the final pictures), and fixed a glaring problem with the media query on the main page.

Fixed back to top button on the contact page (the z-index caused it to sit atop the modal)

Added real website information to my xml file.

No other notes for now. Will continue to revise anything I see in dire need.

Still need the new about-background image file to transform into an .svg.